chore: split websocket runtime module
