+++ test/modules/WebSocket/interceptor.runtime.js
@@ -1,0 +1,11 @@
+import { createInterceptor } from '@mswjs/interceptors'
+import { interceptWebSocket } from '@mswjs/interceptors/lib/interceptors/WebSocket'
+
+const interceptor = createInterceptor({
+  modules: [interceptWebSocket],
+  resolver(event) {
+    window.resolver(event)
+  },
+})
+
+interceptor.apply()
